[CHANGE] moved `nanos()` and `millis()` to nats-base-client as they are referenced by the base library [CI] [BUMP] nats-server
