+++ jetstream/internal_mod.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The NATS Authors
+ * Copyright 2023-2024 The NATS Authors
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -12,13 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export {
-  checkJsError,
-  isFlowControlMsg,
-  isHeartbeatMsg,
-  millis,
-  nanos,
-} from "./jsutil.ts";
+export { checkJsError, isFlowControlMsg, isHeartbeatMsg } from "./jsutil.ts";
 
 export {
   AdvisoryKind,
