+++ jetstream/jsapi_types.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The NATS Authors
+ * Copyright 2023-2024 The NATS Authors
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -14,7 +14,8 @@
  */
 
 import { ApiError, Nanos } from "../nats-base-client/core.ts";
-import { nanos } from "./jsutil.ts";
+
+import { nanos } from "../nats-base-client/util.ts";
 
 export interface ApiPaged {
   total: number;
