+++ jetstream/jsutil.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2023 The NATS Authors
+ * Copyright 2021-2024 The NATS Authors
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -19,7 +19,6 @@ import { MsgImpl } from "../nats-base-client/msg.ts";
 import {
   ErrorCode,
   Msg,
-  Nanos,
   NatsError,
   Publisher,
 } from "../nats-base-client/core.ts";
@@ -87,22 +86,6 @@ export function validName(name = ""): string {
     }
   }
   return "";
-}
-
-/**
- * Converts the specified millis into Nanos
- * @param millis
- */
-export function nanos(millis: number): Nanos {
-  return millis * 1000000;
-}
-
-/**
- * Convert the specified Nanos into millis
- * @param ns
- */
-export function millis(ns: Nanos) {
-  return Math.floor(ns / 1000000);
 }
 
 /**
