+++ jetstream/mod.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The NATS Authors
+ * Copyright 2023-2024 The NATS Authors
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -16,8 +16,6 @@ export {
   checkJsError,
   isFlowControlMsg,
   isHeartbeatMsg,
-  millis,
-  nanos,
 } from "./internal_mod.ts";
 
 export {
