+++ jetstream/tests/jetream409_test.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2023 The NATS Authors
+ * Copyright 2022-2024 The NATS Authors
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -13,9 +13,10 @@
  * limitations under the License.
  */
 
-import { Js409Errors, nanos, setMaxWaitingToFail } from "../jsutil.ts";
+import { Js409Errors, setMaxWaitingToFail } from "../jsutil.ts";
 import {
   deferred,
+  nanos,
   NatsError,
   StringCodec,
 } from "../../nats-base-client/mod.ts";
