+++ jetstream/tests/jetstream_test.ts
@@ -23,7 +23,6 @@ import {
   JsHeaders,
   JsMsg,
   JsMsgCallback,
-  nanos,
   PubAck,
   RepublishHeaders,
   RetentionPolicy,
@@ -41,6 +40,7 @@ import {
   ErrorCode,
   headers,
   JSONCodec,
+  nanos,
   NatsError,
   nuid,
   StringCodec,
