+++ jetstream/tests/jsm_test.ts
@@ -30,6 +30,7 @@ import {
   headers,
   JSONCodec,
   jwtAuthenticator,
+  nanos,
   NatsConnection,
   NatsError,
   nkeys,
@@ -43,7 +44,6 @@ import {
   ConsumerInfo,
   DiscardPolicy,
   Lister,
-  nanos,
   PubAck,
   RetentionPolicy,
   StorageType,
