+++ nats-base-client/mod.ts
@@ -17,7 +17,9 @@ export {
   jwtAuthenticator,
   Match,
   Metric,
+  millis,
   MsgHdrsImpl,
+  nanos,
   NatsError,
   nkeyAuthenticator,
   nkeys,
